Rename WhitelistedItemsIter to ItemTraversal
Also moves it out to its own ir::traversal module, and moves ItemSet to the
ir::item module.
