+++ src/ir/item.rs
@@ -7,11 +7,12 @@ use super::function::Function;
 use super::item_kind::ItemKind;
 use super::module::Module;
 use super::ty::{TemplateDeclaration, Type, TypeKind};
-use super::type_collector::{ItemSet, TypeCollector};
+use super::type_collector::TypeCollector;
 use clang;
 use clang_sys;
 use parse::{ClangItemParser, ClangSubItemParser, ParseError, ParseResult};
 use std::cell::{Cell, RefCell};
+use std::collections::BTreeSet;
 use std::fmt::Write;
 use std::iter;
 
@@ -901,6 +902,9 @@ impl Item {
         }
     }
 }
+
+/// A set of items.
+pub type ItemSet = BTreeSet<ItemId>;
 
 impl TemplateDeclaration for ItemId {
     fn template_params(&self, ctx: &BindgenContext) -> Option<Vec<ItemId>> {
